[NF] Support for Windows Renode with .net platform (#2)
Added support for Renode on Windows. 
A workaround was necessary due to the Renode initializer list compilation failure. 

Also added a stub of the PSM peripheral and fixed some GPIO bugs.

+++ piosim/pio.cpp
@@ -112,7 +112,7 @@ PioSimulator::PioSimulator()
       .read = nullptr,
       .write =
         [this, i](uint32_t data) {
-          program_[i] = data;
+          program_[i] = static_cast<uint16_t>(data);
         },
     };
   }
@@ -236,7 +236,8 @@ uint32_t PioSimulator::execute(uint32_t steps)
 
     if (break_in_next)
     {
-      return steps; // i + 1;
+      return steps;
+      // return i + 1;
     }
   }
